+++ src/components/ToastContainer/Toast/index.tsx
@@ -1,0 +1,47 @@
+import React, {useEffect} from 'react';
+
+import { ToastMessage, useToast } from '../../../hooks/ToastContext'
+import { Container } from './styles';
+import { FiAlertCircle, FiXCircle, FiCheckCircle, FiInfo} from 'react-icons/fi'
+
+interface ToastProps {
+  message: ToastMessage;
+  style: object;
+}
+
+const icons = {
+  info: <FiInfo size={24} />,
+  error: <FiAlertCircle size={24} />,
+  success: <FiCheckCircle size={24} />,
+}
+
+const Toast: React.FC<ToastProps> = ({message, style}) => {
+  const { removeToast } = useToast();
+
+  useEffect(() => {
+      const timer = setTimeout(() => {
+        removeToast(message.id);
+      }, 3000)
+
+      return () => {
+        clearTimeout(timer);
+      }
+  },[removeToast, message.id])
+
+  
+
+  return (
+    <Container type={message.type} hasDescription={!!message.description} style={style} >
+      {icons[message.type]}
+          <div> 
+            <strong>{message.title}</strong>
+            {message.description && <p> {message.description}  </p>}
+            <button onClick={() => removeToast(message.id)} type="button">
+              <FiXCircle size={18} />
+            </button>
+          </div>
+    </Container>
+  )
+}
+
+export default Toast;
